requirements.txt
feedparser==6.0.12
pyyaml==6.0.3
requests==2.33.1
jinja2==3.1.6
google-genai==1.73.1
litellm==1.83.8
python-dotenv==1.0.1
